🐝 split de la bluetooltip dans son propre fichier
